+++ CSK_Module_TCPIPClient/scripts/Communication/TCPIPClient/TCPIPClient_Controller.lua
@@ -18,7 +18,7 @@ tmrTCPIPClient:setPeriodic(false)
 local triggerValue, eventValue -- Selected trigger command + eventName for trigger/event pair via UI
 local selectedTrigger = '' -- Selected trigger/event pair
 
-local eventToForward -- Preset event name to add via UI (see 'addEventToForwardViaUI')
+local eventToForward = '' -- Preset event name to add via UI (see 'addEventToForwardViaUI')
 local selectedEventToForward = '' -- Selected event to forward content to TCP/IP server within UI table
 
 -- Reference to global handle
@@ -151,18 +151,18 @@ local function selectEventToForwardViaUI(selection)
 
   if selection == "" then
     selectedEventToForward = ''
-    _G.logger:info(nameOfModule .. ": Did not find EventToForward")
+    _G.logger:info(nameOfModule .. ": Did not find EventToForward. Is empty")
   else
     local _, pos = string.find(selection, '"EventToForward":"')
     if pos == nil then
-      _G.logger:info(nameOfModule .. ": Did not find EventToForward")
+      _G.logger:info(nameOfModule .. ": Did not find EventToForward. Is nil")
       selectedEventToForward = ''
     else
       pos = tonumber(pos)
       local endPos = string.find(selection, '"', pos+1)
       selectedEventToForward = string.sub(selection, pos+1, endPos-1)
-      if selectedEventToForward == nil then
-        _G.logger:info(nameOfModule .. ": Did not find EventToForward")
+      if ( selectedEventToForward == nil or selectedEventToForward == "" ) then
+        _G.logger:info(nameOfModule .. ": Did not find EventToForward. Is empty or nil")
         selectedEventToForward = ''
       else
         _G.logger:info(nameOfModule .. ": Selected EventToForward: " .. tostring(selectedEventToForward))
@@ -188,10 +188,10 @@ Script.serveFunction("CSK_TCPIPClient.addEventToForwardViaUI", addEventToForward
 
 local function deleteEventToForward(event)
   tcpIpClient_Model.parameters.forwardEvents[event] = nil
-  Script.notifyEvent("TCPIPClient_OnNewEventToForwardList", tcpIpClient_Model.helperFuncs.createJsonList('eventToForward', tcpIpClient_Model.parameters.forwardEvents))
-  _G.logger:info(nameOfModule .. ": Deleted event = " .. tostring(event))
   local suc = Script.deregister(event, tcpIpClient_Model.sendDataViaTCPIP)
+  _G.logger:info(nameOfModule .. ": Deleted event = " .. tostring(event))  
   _G.logger:info(nameOfModule .. ": Success to deregister of event = " .. tostring(suc))
+  Script.notifyEvent("TCPIPClient_OnNewEventToForwardList", tcpIpClient_Model.helperFuncs.createJsonList('eventToForward', tcpIpClient_Model.parameters.forwardEvents))
 end
 Script.serveFunction("CSK_TCPIPClient.deleteEventToForward", deleteEventToForward)